pyproject.toml
[project]
name = "coinmarketcap-news-sentiment-analysis"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
   "psycopg2-binary~=2.9.10",
   "dotenv~=0.9.9",
   "python-dotenv~=1.1.1",
   "selenium~=4.34.2",
   "transformers~=4.53.2",
   "webdriver-manager~=4.0.2",
   "beautifulsoup4~=4.13.4",
   "apache-airflow~=2.4.0",
   "torch~=2.6.0",
   "numpy~=2.3.1",
   "pandas~=2.3.1",
   "loguru~=0.7.3",
   "datasets~=4.0.0",
   "scikit-learn~=1.7.0",
   "pydantic~=2.11.7",
   "fastapi~=0.116.1",
   "tqdm~=4.67.1",
   "accelerate>=0.26.0",
]
